Refactor createFlyout
